+++ libp2p/src/lib.rs
@@ -44,6 +44,8 @@ pub use multiaddr;
 #[doc(inline)]
 pub use libp2p_autonat as autonat;
 #[doc(inline)]
+pub use libp2p_connection_limits as connection_limits;
+#[doc(inline)]
 pub use libp2p_core as core;
 #[cfg(feature = "dcutr")]
 #[doc(inline)]