+++ swarm-test/src/lib.rs
@@ -26,6 +26,7 @@ use libp2p_core::{
     Multiaddr, PeerId, Transport,
 };
 use libp2p_plaintext::PlainText2Config;
+use libp2p_swarm::dial_opts::PeerCondition;
 use libp2p_swarm::{
     dial_opts::DialOpts, AddressScore, NetworkBehaviour, Swarm, SwarmEvent, THandlerErr,
 };
@@ -233,6 +234,7 @@ where
 
         let dial_opts = DialOpts::peer_id(*other.local_peer_id())
             .addresses(external_addresses)
+            .condition(PeerCondition::Always)
             .build();
 
         self.dial(dial_opts).unwrap();