+++ swarm/src/connection/pool.rs
@@ -20,7 +20,6 @@
 // DEALINGS IN THE SOFTWARE.
 #[allow(deprecated)]
 use crate::connection::{Connection, ConnectionId, ConnectionLimit, PendingPoint};
-use crate::connection_limits::ConnectionLimits;
 #[allow(deprecated)]
 use crate::IntoConnectionHandler;
 use crate::{
@@ -306,6 +305,7 @@ where
     THandler: ConnectionHandler,
 {
     /// Creates a new empty `Pool`.
+    #[allow(deprecated)]
     pub fn new(local_id: PeerId, config: PoolConfig, limits: ConnectionLimits) -> Self {
         let (pending_connection_events_tx, pending_connection_events_rx) = mpsc::channel(0);
         let executor = match config.executor {
@@ -871,6 +871,7 @@ impl Drop for NewConnection {
 #[derive(Debug, Clone)]
 pub struct ConnectionCounters {
     /// The effective connection limits.
+    #[allow(deprecated)]
     limits: ConnectionLimits,
     /// The current number of incoming connections.
     pending_incoming: u32,
@@ -883,6 +884,7 @@ pub struct ConnectionCounters {
 }
 
 impl ConnectionCounters {
+    #[allow(deprecated)]
     fn new(limits: ConnectionLimits) -> Self {
         Self {
             limits,
@@ -894,7 +896,8 @@ impl ConnectionCounters {
     }
 
     /// The effective connection limits.
-    #[deprecated(note = "Use the `connection_limits::Behaviour` instead.")]
+    #[deprecated(note = "Use the `libp2p::connection_limits` instead.")]
+    #[allow(deprecated)]
     pub fn limits(&self) -> &ConnectionLimits {
         &self.limits
     }
@@ -1033,6 +1036,65 @@ fn num_peer_established<TInEvent>(
     established.get(&peer).map_or(0, |conns| {
         u32::try_from(conns.len()).expect("Unexpectedly large number of connections for a peer.")
     })
+}
+
+/// The configurable connection limits.
+///
+/// By default no connection limits apply.
+#[derive(Debug, Clone, Default)]
+#[deprecated(note = "Use `libp2p::connectio_limits` instead.")]
+pub struct ConnectionLimits {
+    max_pending_incoming: Option<u32>,
+    max_pending_outgoing: Option<u32>,
+    max_established_incoming: Option<u32>,
+    max_established_outgoing: Option<u32>,
+    max_established_per_peer: Option<u32>,
+    max_established_total: Option<u32>,
+}
+
+#[allow(deprecated)]
+impl ConnectionLimits {
+    /// Configures the maximum number of concurrently incoming connections being established.
+    pub fn with_max_pending_incoming(mut self, limit: Option<u32>) -> Self {
+        self.max_pending_incoming = limit;
+        self
+    }
+
+    /// Configures the maximum number of concurrently outgoing connections being established.
+    pub fn with_max_pending_outgoing(mut self, limit: Option<u32>) -> Self {
+        self.max_pending_outgoing = limit;
+        self
+    }
+
+    /// Configures the maximum number of concurrent established inbound connections.
+    pub fn with_max_established_incoming(mut self, limit: Option<u32>) -> Self {
+        self.max_established_incoming = limit;
+        self
+    }
+
+    /// Configures the maximum number of concurrent established outbound connections.
+    pub fn with_max_established_outgoing(mut self, limit: Option<u32>) -> Self {
+        self.max_established_outgoing = limit;
+        self
+    }
+
+    /// Configures the maximum number of concurrent established connections (both
+    /// inbound and outbound).
+    ///
+    /// Note: This should be used in conjunction with
+    /// [`ConnectionLimits::with_max_established_incoming`] to prevent possible
+    /// eclipse attacks (all connections being inbound).
+    pub fn with_max_established(mut self, limit: Option<u32>) -> Self {
+        self.max_established_total = limit;
+        self
+    }
+
+    /// Configures the maximum number of concurrent established connections per peer,
+    /// regardless of direction (incoming or outgoing).
+    pub fn with_max_established_per_peer(mut self, limit: Option<u32>) -> Self {
+        self.max_established_per_peer = limit;
+        self
+    }
 }
 
 /// Configuration options when creating a [`Pool`].